Refactor: Refactor statistics widgets' layout
