+++ lib/src/views/home/home_seg3_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import 'widgets/daily_summary/daily_summary_widget.dart';
+import 'widgets/daily_summary/daily_statistics_widget.dart';
 
 //요약
 class Seg3Screen extends StatelessWidget {
@@ -8,6 +8,6 @@ class Seg3Screen extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return DailySummaryWidget();
+    return DailyStatisticsWidget();
   }
 }
