+++ lib/src/views/statistics/widgets/pie_charts/study_pie_chart_widget.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_screenutil/flutter_screenutil.dart';
 
 import '../../../../../core/enum/mxnRate.dart';
 import '../../../../../core/widgets/mxnContainer.dart';
@@ -25,33 +24,20 @@ class StudyPieChartWidget extends StatelessWidget {
       MxN_child: LayoutBuilder(builder: (context, constraints) {
         return Container(
           color: Colors.white,
-          child: Stack(
+          padding: const EdgeInsets.all(8),
+          child: Column(
+            mainAxisAlignment: MainAxisAlignment.spaceBetween,
+            crossAxisAlignment: CrossAxisAlignment.start,
             children: [
-              Positioned(
-                top: 10.w,
-                left: 13.w,
-                child: Text(
-                  "과목별 비율",
-                  style: TextStyle(
-                    fontSize: 14.sp,
-                    fontWeight: FontWeight.bold,
-                  ),
-                ),
-              ),
-              Align(
-                alignment: Alignment.bottomCenter,
-                child: SizedBox(
-                  width: constraints.maxWidth,
-                  height: constraints.maxHeight * 0.7,
-                  child: Padding(
-                    padding: const EdgeInsets.all(8.0),
-                    child: StudyPieChart(
-                      subjectTitleList: subjectTitleList,
-                      studyDurationList: studyDurationList,
-                      subjectColorList: subjectColorList,
-                      totalStudyDuration: totalStudyDuration,
-                    ),
-                  ),
+              Text("과목별 비율"),
+              SizedBox(
+                width: constraints.maxWidth,
+                height: constraints.maxHeight * 0.7,
+                child: StudyPieChart(
+                  subjectTitleList: subjectTitleList,
+                  studyDurationList: studyDurationList,
+                  subjectColorList: subjectColorList,
+                  totalStudyDuration: totalStudyDuration,
                 ),
               )
             ],
